% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creepr.R
\name{creepr}
\alias{creepr}
\title{Play a creepy noise}
\usage{
creepr(noise)
}
\arguments{
\item{noise}{The chosen noise to play. Must be one of
<insert list of noises>
If no noise is chosen, a random noise is played.}
}
\description{
This function plays a creepy noise
}
